+++ CiscoAXL.py
@@ -200,6 +200,117 @@ def AssociateE911Location(conn, LocationUUID, AbbrevCluster):
     except Exception:
         return False
 
+def BuildCMRGs(conn, AbbrevCluster, groupNum):
+    try:
+        for groupLetter in ["A","B"]:
+            resp = conn.addCallManagerGroup(
+                callManagerGroup = {
+                    'name' : f"{AbbrevCluster}_CMRG_{groupNum}{groupLetter}",
+                    'members' : {
+                        'member' : {
+                            'callManagerName' : 'CM_hq-cucm-pub',
+                            'priority' : 1
+                        }
+                    }      
+                }
+            )
+        cmrg_uuid = resp['return'].strip('{}').lower()
+        
+        return True, cmrg_uuid
+    except Fault as err:
+        return False, err
+    except Exception as err:
+        return False, err
+
+# NEEDS CLEANED UP IF WE ARE GOING TO USE FOR MORE THAN STAGING
+def BuildDTGroups(conn):
+    base_DateTime_List = {'Eastern':'America/New_York','Central':'America/Chicago','Mountain':'America/Denver','Arizona':'America/Phoenix','Pacific':'America/Los_Angeles'}
+    try:
+        for key in base_DateTime_List:
+            resp = conn.addDateTimeGroup(
+                dateTimeGroup = {
+                    'name' : f"{key}",
+                    'timeZone' : f"{base_DateTime_List[key]}",
+                    'separator' : '/',
+                    'dateformat' : 'D/M/Y',
+                    'timeFormat' : '12-hour'
+                }      
+            )
+        dtGroup_uuid = resp['return'].strip('{}').lower()
+        
+        return True, dtGroup_uuid
+    except Fault as err:
+        return False, err
+    except Exception as err:
+        return False, err
+
+# NEED TO BUILD OUT FOR xCODERS and CFBs, MEMBER CURRENTLY HARDCODED FOR LAB 
+def BuildMRGs(conn, AbbrevCluster, GroupNum):
+    try:
+        resp = conn.addMediaResourceGroup(
+            mediaResourceGroup = {
+                'name' : f"{AbbrevCluster}_Hardware_MRG_{GroupNum}",
+                'description' : f"{AbbrevCluster}_Hardware_MRG_{GroupNum}",
+                'multicast' : 'f',
+                'members' : {
+                    'member' : {
+                        'deviceName' : 'ANN_2' 
+                    }   
+                }
+            }
+        )      
+        mrg_uuid = resp['return'].strip('{}').lower()
+
+        return True, mrg_uuid
+    except Fault as err:
+        return False, err
+    except Exception as err:
+        return False, err
+        
+def BuildMRGLs(conn, AbbrevCluster):
+    try:
+        resp = conn.addMediaResourceList(
+            mediaResourceList = {
+                'name' : f"RemoteSite_{AbbrevCluster}_MRGL",
+                'members' : {
+                    'member' : {
+                        'mediaResourceGroupName' : f"{AbbrevCluster}_Hardware_MRG_1",
+                        'order' : 1
+                    }   
+                }
+            }      
+        )        
+        mrgl_uuid = resp['return'].strip('{}').lower()
+        
+        return True, mrgl_uuid
+    except Fault as err:
+        return False, err
+    except Exception as err:
+        return False, err
+
+def BuildRouteGroups(conn, AbbrevCluster, Carrier):
+    try:
+        resp = conn.addRouteGroup(
+            routeGroup = {
+                'name' : f"SBC_{AbbrevCluster}_{Carrier}_RG",
+                'distributionAlgorithm' : "Circular",
+                'members' : {
+                    'member' : {
+                        'deviceSelectionOrder' : 1,
+                        'deviceName' : "SIPTrunktoCUP",
+                        'port' : 1
+                    }   
+                }
+            }      
+        )        
+        routegroup_uuid = resp['return'].strip('{}').lower()
+        
+        return True, routegroup_uuid
+    except Fault as err:
+        return False, err
+    except Exception as err:
+        return False, err
+
 def BuildDevicePool(conn, SiteCode, AbbrevCluster, CMRG, TZ, StandardLRG = None):
     try: 
         resp = conn.addDevicePool(
@@ -242,4 +353,25 @@ def setStandardLRG(conn, DevicePoolUUID, RouteGroup):
     
         return True, ""
     except Fault as err:
+        return False, err
+
+def BuildPartitions(conn, SiteCode, AbbrevCluster):
+    if SiteCode == 'STAGING':
+        partitionNames = {f'{AbbrevCluster}_DN_PT':f'{AbbrevCluster} DN Partition',f'{AbbrevCluster}_Outbound_PT':f'{AbbrevCluster} Outbound Partition',f'E911_{AbbrevCluster}_Hunt_PT':f'{AbbrevCluster} 911 Hunt Partition',f'{AbbrevCluster}_CMService_PT':f'{AbbrevCluster} Service Partition'}
+    else:
+        partitionNames = {f'{SiteCode}_{AbbrevCluster}_Trans_PT':f'{SiteCode} Translation Partition',f'{SiteCode}_{AbbrevCluster}_Outbound_PT':f'{SiteCode} Outbound Partition',f'{SiteCode}_{AbbrevCluster}_Park_PT':f'{SiteCode} Park Partition'}
+    try:
+        for partition in partitionNames:
+            resp = conn.addRoutePartition(
+                routePartition = {
+                    'name' : f"{partition}",
+                    'description' : f"{partitionNames[partition]}"
+                }
+            )
+        partition_uuid = resp['return'].strip('{}').lower()
+            
+        return True, partition_uuid
+    except Fault as err:
+        return False, err
+    except Exception as err:
         return False, err
